Multi-user support
Added new config option:
users: [
  {
    "id": 42,
    "username": "guest",
    "password": "s3cr3t", // or a password hash
    "admin": false
  }
]
There is also the old rpc user, which has (virtually) id -1 and
it's an admin. The id is used to keep track of the owner of a torrent.
Changing the ids will make the users loose their torrents. Note that the
admins are still able to see all the torrents.

An admin can do anything as before, a non-admin user has restricted
privileges:

- Sees only his torrents (each torrent is bound to a specific user)
- Cannot interact (sort/start/pause/...) other user's torrents
- Cannot choose an absolute path, only a path relative to
  download_dir/username
- Cannot stop the server
- Cannot change path of a torrent
- Cannot change server settings

As the old rpc password, also the ones stored in the users list will be
hashed and overwritten in the settings.json file.

In the web-ui there is an extra "Owner" field in the details view, it
shows the username of the user that created that torrent.